Add a protocl to the datasources
I had to modify the lookpback-connector-mongodb dependency to hardcode
the mongodb protocol because it defaults to `mongodb` but we need to use
`mongodb+srv` for resolving a server by its domain name. The better
approach, in my opinion, is to add a protocol field to the datasources
and pass that in when creating the mongodb client.